Moved otp functions to dedicated file, small improvements
