+++ src/gauth.py
@@ -2,62 +2,19 @@
 
 import sys
 import os
-import hmac
-import base64
-import struct
-import hashlib
-import time
 import ConfigParser
 import os.path
-import math
+import time
 
 import alfred
-
-
-def get_hotp_token(key, intervals_no):
-    msg = struct.pack(">Q", intervals_no)
-    h = hmac.new(key, msg, hashlib.sha1).digest()
-    o = ord(h[19]) & 15
-    h = (struct.unpack(">I", h[o:o + 4])[0] & 0x7fffffff) % 1000000
-    return h
-
-
-def get_totp_token(key):
-    token = get_hotp_token(key, intervals_no=int(time.time()) // 30)
-    return str(token).zfill(6)
-
-
-def get_totp_time_remaining():
-    return int(30 - (time.time() % 30))
-
-
-def is_secret_valid(secret):
-    try:
-        secret = secret.replace(' ', '')
-        secret = secret.ljust(int(math.ceil(len(secret) / 16.0) * 16), '=')
-        key = base64.b32decode(secret, casefold=True)
-        get_totp_token(key)
-    except:
-        return False
-
-    return True
-
-
-def get_hotp_key(key=None, secret=None, hexkey=None):
-    if hexkey:
-        key = hexkey.decode('hex')
-    if secret:
-        secret = secret.replace(' ', '')
-        secret = secret.ljust(int(math.ceil(len(secret) / 16.0) * 16), '=')
-        key = base64.b32decode(secret, casefold=True)
-    return key
+import otp
 
 
 class AlfredWorkflow(object):
     _reserved_words = []
 
     def write_text(self, text):
-        return alfred.write(text)
+        print(text)
 
     def write_item(self, item):
         return self.write_items([item])
@@ -112,17 +69,19 @@ class AlfredGAuth(AlfredWorkflow):
     _reserved_words = ['add', 'update', 'remove']
 
     def __init__(self, config_file='~/.gauth', max_results=20):
+        self.max_results = max_results
+
         self._config_file = config_file
         self.config_file = os.path.expanduser(self._config_file)
-        self.max_results = max_results
-        self._config = None
+        self.config = ConfigParser.RawConfigParser()
+        self.config.read(self.config_file)
 
         # If the configuration file doesn't exist, create an empty one
         if not os.path.isfile(self.config_file):
             self.create_config()
 
         try:
-            if not self.config.sections() and action != 'add':
+            if not self.config.sections():
                 # If the configuration file is empty,
                 # tell the user to add secrets to it
                 self.write_item(self.config_file_is_empty_item())
@@ -133,15 +92,8 @@ class AlfredGAuth(AlfredWorkflow):
                                        exception=e)
             self.write_item(item)
 
-    @property
-    def config(self):
-        if not self._config:
-            self._config = ConfigParser.RawConfigParser()
-            self._config.read(os.path.expanduser(self.config_file))
-        return self._config
-
     def create_config(self):
-        with open(os.path.expanduser(self.config_file), 'w') as f:
+        with open(self.config_file, 'w') as f:
             f.write(self._config_file_initial_content)
             f.close()
 
@@ -161,8 +113,8 @@ class AlfredGAuth(AlfredWorkflow):
         except:
             hexkey = None
 
-        key = get_hotp_key(secret=secret, key=key, hexkey=hexkey)
-        return get_totp_token(key)
+        key = otp.get_hotp_key(secret=secret, key=key, hexkey=hexkey)
+        return otp.get_totp_token(key)
 
     def config_list_accounts(self):
         return self.config.sections()
@@ -178,7 +130,7 @@ class AlfredGAuth(AlfredWorkflow):
     def time_remaining_item(self):
         # The uid for the remaining time will be the current time,
         # so it will appears always at the last position in the list
-        time_remaining = get_totp_time_remaining()
+        time_remaining = otp.get_totp_time_remaining()
         return alfred.Item({u'uid': time.time(), u'arg': '', u'ignore': 'yes'},
                            'Time Remaining: {}s'.format(time_remaining),
                            None, 'time.png')
@@ -187,7 +139,7 @@ class AlfredGAuth(AlfredWorkflow):
         return self.warning_item(title='GAuth is not yet configured',
                                  message='You must add your secrets to '
                                  'the {} file (see documentation)'
-                                 .format(self.config_file))
+                                 .format(self._config_file))
 
     def search_by_account_iter(self, query):
         if self.is_command(query):
@@ -208,11 +160,10 @@ class AlfredGAuth(AlfredWorkflow):
                                     'There is no account matching "{}" '
                                     'on your configuration file '
                                     '({})'.format(query,
-                                                  self.config_file))
+                                                  self._config_file))
 
     def add_account(self, account, secret):
-
-        if not is_secret_valid(secret):
+        if not otp.is_otp_secret_valid(secret):
             return "Invalid secret:\n[{0}]".format(secret)
 
         config_file = open(self.config_file, 'r+')
@@ -238,7 +189,8 @@ class AlfredGAuth(AlfredWorkflow):
         except ValueError:
             return self.write_text('Invalid arguments!\n'
                                    'Please enter: account, secret.')
-        self.write_text(self.add_account(account, secret))
+
+        return self.write_text(self.add_account(account, secret))
 
 
 def main(action, query):
